refactor(reservation): decouple UI from services and split resolveCounterparty into SRP helpers
- Moved ??????????????? src/lib/reservation/resolveCounterparty.ts????????? service -> hook -> component ????????UI ????????? services??
- ??? esolveCounterparty ?????????????RP???????????resolveCounterpartyId????????ID ????????? esolveCounterpartyProfile?????????????????????/??????????????????????????code smell??
- ???????????????????mutations ???????????????????????????????????????????????
